pyproject.toml
[project]
name = "dailymovie"
version = "0.1.0"
description = "DailyMovie - 每日电影推荐"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "httpx>=0.26.0",
    "sqlalchemy>=2.0.25",
    "aiosqlite>=0.19.0",
    "python-dotenv>=1.0.0",
    "pillow>=12.1.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.26.0",
    "black>=23.0.0",
    "ruff>=0.2.0",
]

[tool.uv]

[tool.hatch.build.targets.wheel]
packages = ["backend"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
